Refactor Gate symbol conversion
