+++ src/core/exchange_utils.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <string>
+
+std::string to_gate_symbol(const std::string &symbol);
+
